add in start script and move modal start
